Basic function worked

+++ src/sample/controller/GameController.java
@@ -28,17 +28,21 @@ public class GameController {
     @FXML private TableColumn<Game,Integer> gameIdColumn;
     @FXML private TableColumn<Game,String> gameNameColumn;
     @FXML private TableColumn<Game,String> gameDevelopColumn;
-    @FXML private TableColumn<Game, java.sql.Date> gameCreateDateColumn;
     @FXML private TableColumn<Game,Double> gameCostColumn;
 
     // Search a game
     @FXML
-    private void searchGame (ActionEvent actionEvent) throws ClassNotFoundException,SQLException{
+    private void searchGame () throws ClassNotFoundException,SQLException{
         try {
             // Get Game information
+            if (gameIdText.getText().equals("")) {
+                resultArea.setText("Search INFO:\nPlease select Game!");
+                loadTableFromDataBase();
+                return;
+            }
             Game game = GameDAO.searchGame(gameIdText.getText());
             // Populate Game on TableView and Display on TextArea
-            populateAndShowGame(game);
+            loadGameWithInfo(game);
         }catch (SQLException e) {
             e.printStackTrace();
             resultArea.setText("Error accurred while getting employee information from DB.\n" + e);
@@ -47,12 +51,12 @@ public class GameController {
     }
     // Search all games
     @FXML
-    private void searchGames (ActionEvent actionEvent) throws SQLException, ClassNotFoundException {
+    private void loadTableFromDataBase() throws SQLException, ClassNotFoundException {
         try {
             // Get all games information
-            ObservableList<Game> gameData = GameDAO.searchGames();
+            ObservableList<Game> gameData = GameDAO.findAllGames();
             // Populate Games on TableView
-            populateGame((Game) gameData);
+            gamesTable.setItems(gameData);
         }catch (SQLException e) {
             System.out.println("Error accured while getting games information from DB.\n");
             throw e;
@@ -61,16 +65,15 @@ public class GameController {
     // Initializing the controller class.
     // This method is automatically called after the fxml file has been loaded.
     @FXML
-    private void initialize () {
+    private void initialize () throws SQLException ,  ClassNotFoundException{
         gameIdColumn.setCellValueFactory(cellData -> cellData.getValue().game_idProperty().asObject());
         gameNameColumn.setCellValueFactory(cellData -> cellData.getValue().nameProperty());
         gameDevelopColumn.setCellValueFactory(cellData -> cellData.getValue().developProperty());
-        gameCreateDateColumn.setCellValueFactory(cellData -> cellData.getValue().create_dateProperty());
         gameCostColumn.setCellValueFactory(cellData -> cellData.getValue().costProperty().asObject());
+        loadTableFromDataBase();
     }
-    // Populate Game
     @FXML
-    private void populateGame (Game game) throws ClassNotFoundException {
+    private void loadGameToTable (Game game) throws ClassNotFoundException {
         // Declare and ObservableList for table view
         ObservableList<Game> gamesData = FXCollections.observableArrayList();
         // Add game to ObservableList
@@ -88,24 +91,25 @@ public class GameController {
     }
     // Populate Game for TableView and Display Game on TextArea
     @FXML
-    private void populateAndShowGame (Game game) throws ClassNotFoundException{
+    private void loadGameWithInfo(Game game) throws ClassNotFoundException{
         if (game != null){
-            populateGame(game);
+            loadGameToTable(game);
             setGameInfoToTextArea(game);
         }else {
             resultArea.setText("This Game does not exist !\n");
+            gamesTable.setItems(null);
         }
     }
-    // Populate Game for TableView
     @FXML
-    private void populateGames (ObservableList<Game> gamesData) throws ClassNotFoundException{
-        // Set items to the gamesTable
-        gamesTable.setItems(gamesData);
-    }
-    // Update game's cost with cost with the cost is written on newCostText field
-    @FXML
-    private void updateGameCost (ActionEvent actionEvent) throws SQLException, ClassNotFoundException {
+    private void updateGameCost () throws SQLException, ClassNotFoundException {
         try {
+            if (gameIdText.getText().equals("")) {
+                resultArea.setText("Update INFO\nPlease select Game!");
+                return;
+            }else if (newCostText.getText().equals("")){
+                resultArea.setText("Please insert new Cost");
+                return;
+            }
             GameDAO.updateGameCost(gameIdText.getText(),Double.valueOf(newCostText.getText()));
             resultArea.setText("Cost has been updated for, game id: " + gameIdText.getText() + "\n");
         }catch (SQLException e){
@@ -113,23 +117,31 @@ public class GameController {
             throw e;
         }
     }
-    // Insert a game to the DB
     @FXML
     private void insertGame (ActionEvent actionEvent) throws SQLException, ClassNotFoundException {
         try {
+            if (nameText.getText().equals("") || developText.getText().equals("") || costText.getText().equals("")){
+                resultArea.setText("Please check information about game !");
+                return;
+            }
             GameDAO.insertGame(nameText.getText(),developText.getText(),Double.valueOf(costText.getText()));
-            resultArea.setText("Game inserted! \n");
+            loadTableFromDataBase();
+            resultArea.setText("Insert INFO\nGame inserted!\n" +
+                    "Game: " + nameText.getText());
         }catch (SQLException e) {
             resultArea.setText("Problem accurred while inserting game " + e);
-            throw e;
         }
     }
     //Delete a game with a given game id from DB
     @FXML
     private void deleteGame (ActionEvent actionEvent) throws SQLException,ClassNotFoundException{
         try {
+            if (gameIdText.getText().equals("")) {
+                resultArea.setText("Delete INFO:\nPlease select Game!");
+                return;
+            }
             GameDAO.deleteGameWithId(gameIdText.getText());
-            resultArea.setText("Game deleted! Game id: " + gameIdText.getText() + "\n");
+            resultArea.setText("Game deleted! Game: " + gameIdText.getText() + "\n");
         }catch (SQLException e) {
             resultArea.setText("Problem accured while deleting game " + e);
             throw e;
